refactor(server-remote-standalone): rename server.ts to app.ts
Aligns with the cloudflare package's naming convention where the main
Hono app module is app.ts.
