+++ packages/server-remote-standalone/src/index.ts
@@ -1,3 +1,3 @@
-export type { StandaloneAuthConfig, StandaloneHubConfig } from './server';
-export { createRemoteHub } from './server';
+export type { StandaloneAuthConfig, StandaloneHubConfig } from './app';
+export { createRemoteHub } from './app';
 export { BunSqliteUpdateLog } from './storage';
