+++ packages/server-remote-standalone/src/start.ts
@@ -8,8 +8,8 @@
  */
 
 import { mkdirSync } from 'node:fs';
-import type { StandaloneAuthConfig } from './server';
-import { createRemoteHub } from './server';
+import type { StandaloneAuthConfig } from './app';
+import { createRemoteHub } from './app';
 
 function resolveAuthConfig(): StandaloneAuthConfig {
 	if (process.env.AUTH_TOKEN) {
